Split babel.js in smaller files
