+++ package.json
@@ -57,6 +57,7 @@
   "dependencies": {
     "babel-plugin-syntax-jsx": "6.18.0",
     "babel-traverse": "6.21.0",
+    "babel-types": "6.23.0",
     "babylon": "6.14.1",
     "convert-source-map": "1.3.0",
     "escape-string-regexp": "1.0.5",
