+++ src/_constants.js
@@ -1,0 +1,6 @@
+export const STYLE_ATTRIBUTE = 'jsx'
+export const GLOBAL_ATTRIBUTE = 'global'
+export const STYLE_COMPONENT_ID = 'styleId'
+export const MARKUP_ATTRIBUTE = 'data-jsx'
+export const STYLE_COMPONENT = '_JSXStyle'
+export const STYLE_COMPONENT_CSS = 'css'
